【柴方晨】投放列表，投放详情和其他一些细节处理

+++ card-coupon/src/main/java/com/joycity/joyclub/card_coupon/service/impl/CardCouponLaunchServiceImpl.java
@@ -5,14 +5,19 @@ import static com.joycity.joyclub.commons.constant.ResultCode.LAUNCH_NUM_EXCEED_
 import java.util.Date;
 import java.util.List;
 
+import com.joycity.joyclub.card_coupon.constant.CouponCodeUseStatus;
 import com.joycity.joyclub.card_coupon.constant.CouponLaunchReviewStatus;
 import com.joycity.joyclub.card_coupon.constant.CouponLaunchType;
 import com.joycity.joyclub.card_coupon.constant.CouponType;
+import com.joycity.joyclub.card_coupon.mapper.CardCouponCodeMapper;
 import com.joycity.joyclub.card_coupon.mapper.CardCouponLaunchMapper;
 import com.joycity.joyclub.card_coupon.mapper.CardCouponMapper;
+import com.joycity.joyclub.card_coupon.mapper.CardCouponTriggerScopeMapper;
 import com.joycity.joyclub.card_coupon.modal.CreateCouponInfo;
+import com.joycity.joyclub.card_coupon.modal.CreateCouponLaunchInfo;
 import com.joycity.joyclub.card_coupon.modal.ShowCouponLaunchInfo;
 import com.joycity.joyclub.card_coupon.modal.generated.CardCouponLaunch;
+import com.joycity.joyclub.card_coupon.modal.generated.CardCouponTriggerScope;
 import com.joycity.joyclub.card_coupon.service.CardCouponLaunchService;
 import com.joycity.joyclub.commons.AbstractGetListData;
 import com.joycity.joyclub.commons.constant.ResultCode;
@@ -36,11 +41,22 @@ public class CardCouponLaunchServiceImpl implements CardCouponLaunchService {
     private CardCouponLaunchMapper launchMapper;
     @Autowired
     private CardCouponMapper cardCouponMapper;
+    @Autowired
+    private CardCouponTriggerScopeMapper cardCouponTriggerScopeMapper;
+    @Autowired
+    private CardCouponCodeMapper cardCouponCodeMapper;
+
 
     @Override
-    public ResultData createCardCouponLaunch(CardCouponLaunch launch) {
+    public ResultData createCardCouponLaunch(CreateCouponLaunchInfo launch) {
         checkLaunchNum(launch);
         launchMapper.insertSelective(launch);
+        if (CollectionUtils.isNotEmpty(launch.getCouponTriggerScopes())) {
+            for (CardCouponTriggerScope triggerScope : launch.getCouponTriggerScopes()) {
+                triggerScope.setLaunchId(launch.getId());
+                cardCouponTriggerScopeMapper.insertSelective(triggerScope);
+            }
+        }
         return new ResultData(new CreateResult(launch.getId()));
     }
 
@@ -74,6 +90,13 @@ public class CardCouponLaunchServiceImpl implements CardCouponLaunchService {
             }
         }
         return data;
+    }
+
+    @Override
+    public ResultData getCardCouponLaunchById(Long id) {
+        CreateCouponLaunchInfo createCouponLaunchInfo = launchMapper.selectCouponLaunchInfoById(id);
+        launchResult(createCouponLaunchInfo);
+        return new ResultData(createCouponLaunchInfo) ;
     }
 
     @Override
@@ -122,24 +145,42 @@ public class CardCouponLaunchServiceImpl implements CardCouponLaunchService {
 
     @Override
     public ResultData deleteCardCouponLaunch(Long id) {
+        CardCouponLaunch launchDb = launchMapper.selectByPrimaryKey(id);
+        if (!launchDb.getReviewStatus().equals(CouponLaunchReviewStatus.STATUS_NOT_REVIEW)) {
+            throw new BusinessException(ResultCode.LAUNCH_ERROR, "只有未审核，才可以删除");
+        }
         return new ResultData(new UpdateResult(launchMapper.deleteCardCouponLaunchById(id)));
     }
 
     private void checkLaunchNum(CardCouponLaunch launch){
         CreateCouponInfo couponInfo = cardCouponMapper.selectCardCouponById(launch.getCouponId());
-        int launchSum = launchMapper.selectlaunchNumByCouponId(couponInfo.getId());
-        int remainNum = couponInfo.getNum() - launchSum;
-        if (remainNum < launch.getLaunchNum()) {
-            throw new BusinessException(LAUNCH_NUM_EXCEED_COUPON_NUM, "投放数量超过剩余发行量");
-        }
-        if (couponInfo.getType().equals(CouponType.DEDUCTION_COUPON)) {
-            if (launch.getType().equals(CouponLaunchType.CONDITION_LAUNCH)) {
-                throw new BusinessException(LAUNCH_NUM_EXCEED_COUPON_NUM, "满减券只能选择批量投放和线上投放");
+        if (couponInfo != null) {
+            int launchSum = launchMapper.selectlaunchNumByCouponId(couponInfo.getId());
+            int remainNum = couponInfo.getNum() - launchSum;
+            if (remainNum < launch.getLaunchNum()) {
+                throw new BusinessException(LAUNCH_NUM_EXCEED_COUPON_NUM, "投放数量超过剩余发行量");
             }
-        } else if (couponInfo.getType().equals(CouponType.CASH_COUPON)) {
-            if (!launch.getType().equals(CouponLaunchType.CONDITION_LAUNCH)) {
-                throw new BusinessException(LAUNCH_NUM_EXCEED_COUPON_NUM, "代金券只能选择条件投放");
+            if (couponInfo.getType().equals(CouponType.DEDUCTION_COUPON)) {
+                if (launch.getType().equals(CouponLaunchType.CONDITION_LAUNCH)) {
+                    throw new BusinessException(LAUNCH_NUM_EXCEED_COUPON_NUM, "满减券只能选择批量投放和线上投放");
+                }
+            } else if (couponInfo.getType().equals(CouponType.CASH_COUPON)) {
+                if (!launch.getType().equals(CouponLaunchType.CONDITION_LAUNCH)) {
+                    throw new BusinessException(LAUNCH_NUM_EXCEED_COUPON_NUM, "代金券只能选择条件投放");
+                }
             }
         }
     }
+
+    /**
+     * 添加投放结果
+     * @param info
+     */
+    private void launchResult(CreateCouponLaunchInfo info){
+        int receiveNum = cardCouponCodeMapper.countByLaunchId(info.getId());
+        info.setReceiveNum(receiveNum);
+        int usedNum = cardCouponCodeMapper.countByLaunchIdAndUseStatus(info.getId(), CouponCodeUseStatus.USED);
+        info.setUsedNum(usedNum);
+    }
+
 }
